+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1320,7 +1320,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T09:41:08.844Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T13:43:09.609Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1372,7 +1372,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T09:41:08.843Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T13:43:09.608Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1424,7 +1424,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[Source] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T09:41:08.843Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2025-11-03T13:43:09.607Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1498,5 +1498,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-03T09:40:57.201Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T13:42:58.298Z")
     r"""The date and time the resource was last updated"""