+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K94HF0G90X6BR68TK9VQPNVQ"
+        "01K94ZA4Y6YHWZDR98E1VJWA55"
     )
     r"""The unique identifier of the data source"""