+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-03T09:40:57.521Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T13:42:58.679Z")
     r"""The date and time the resource was last updated"""