+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-03T09:40:57.201Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-03T13:42:58.298Z")
     r"""The date and time the resource was last updated"""