+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K94HF0GCQNR9HKC8ZG4KGRM9"
+        "01K94ZA4YBV9MW6B9HYPWB6DHT"
     )
     r"""The unique identifier of the data source"""